requirements.txt
google-genai==1.21.0
langgraph==0.4.5
python-dotenv==1.0.1
openai==1.88.0
grandalf==0.8
langchain-google-genai
langchain-core
